+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Drop obsolete `GetSpecifiedInstance` method from `IReadOnlyNamedOptions` and `IWritableNamedOptions` (use `GetInstance` instead)
 - Drop obsolete `FormatProviderBase.LoadConfiguration<T>(WritableOptionsConfiguration<T>)` overloads (use `LoadConfiguration(Type, IWritableOptionsConfiguration)` instead)
 - Drop obsolete `FormatProviderBase.SaveAsync<T>(T, WritableOptionsConfiguration<T>, CancellationToken)` overload (use `SaveAsync<T>(T, IWritableOptionsConfiguration, CancellationToken)` instead)
+- Drop `OnChangeThrottleMs` property (use `OnChangeThrottle` with `TimeSpan` instead)
 
 ## [0.5.0] - 2026-07-11